\name{MultiOmics-class}
\docType{class}
\alias{MultiOmics}
\alias{MultiOmics-class}
\alias{summary,MultiOmics-method}
\alias{[,MultiOmics,ANY,ANY,ANY-method}
\alias{plot,MultiOmics-method}
\alias{validMultiOmics}
\alias{prepareMultiOmics}

\title{Class \code{"MultiOmics"}}
\description{
  The \code{MultiOmics} object class ....
  The \code{prepareMultiOmics} function returns a new object of \code{MultiOmics} class.
}
\section{Objects from the Class}{
  Objects should be defined using the \code{prepareMultiOmics} constructor. In
  the simplest case, you simply ....
}
\usage{
prepareMultiOmics(datalist, outcome)
\S4method{summary}{MultiOmics}(object, \dots){}
\S4method{plot}{MultiOmics,missing}(x, y, \dots){}
#\S4method{[]}{x, i, j, ..., drop = FALSE}{}
}
\arguments{
  \item{datalist}{....}
  \item{outcome}{....}
}
\section{Slots}{
  \describe{
      \item{\code{data}:}{...}
      \item{\code{outcome}:}{...}
  }
}
\section{Methods}{
  \describe{
      \item{\code{[]}:}{....}
      \item{\code{\link{plot}}:}{....}
      \item{\code{\link{summary}}:}{....}
  }
}
\value{
The \code{prepareMultiOmics} function returns a new object of the \code{MultiOmics} class.
}
\author{Kevin R. Coombes <krc@silicovore.com>}
\note{
}
\seealso{}
\examples{
}
\keyword{  }
