% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIA.R
\name{aria}
\alias{aria}
\title{Adjusted Rand Index Analysis}
\usage{
aria(
  points,
  number_of_trials = 100,
  clusterpersistence = NULL,
  clusterradius = NULL,
  outputall = FALSE
)
}
\arguments{
\item{points}{Point pattern.}

\item{clusterradius}{Radius for determining neighbourhood.}

\item{clusterperistence}{Persistence threshold for ToMATo.}
}
\value{
Adjusted rand index and associated p value.
}
\description{
Outputs the ARI value and the estimated p-value associated with it for a given data set.
}
