% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{bindatum}
\alias{bindatum}
\title{Bin Datum function used in MPP Histogram}
\usage{
bindatum(i, y_wid, GP_x, GP_y, pts)
}
\arguments{
\item{i}{Bin number.}

\item{y_wid}{Width of bin in y direction.}

\item{GP_x}{GP in x.}

\item{GP_y}{GP in y.}

\item{pts}{Points.}
}
\value{
Binned data.
}
\description{
Bins data points together. Dev use only.
}
