% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLASMA.R
\name{bulkanalyse}
\alias{bulkanalyse}
\title{Bulk Analyse}
\usage{
bulkanalyse(cloud, wholeradius, wholedeviance, split, minpoints)
}
\arguments{
\item{cloud}{Point pattern or list of point patterns.}
}
\value{
One set of results from P-Check, JOSEPH and MPP Histogram.
}
\description{
Used in PLASMA.
}
