% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{calcinter}
\alias{calcinter}
\title{Calculate intersections function used in Inform Persistence}
\usage{
calcinter(m1, m2, s1, s2, c1, c2)
}
\arguments{
\item{m1}{Mean (peak) of first Gaussian.}

\item{m2}{Mean (peak) of second Gaussian.}

\item{s1}{Standard deviation of first Gaussian.}

\item{s2}{Standard deviation of second Gaussian.}

\item{c1}{Coefficient of first Gaussian.}

\item{c2}{Coefficient of second Gaussian.}
}
\value{
x value of point of intersection.
}
\description{
Calculates the point of intersection between two Gaussians.
}
