% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{calculate_L_values}
\alias{calculate_L_values}
\title{Calculate Getis L Values function used in MPP Simulate.}
\usage{
calculate_L_values(
  distance_matrix,
  number_of_points,
  domain_size_x,
  domain_size_y,
  search_radius
)
}
\arguments{
\item{distance_matrix}{Distance matrix of point cloud.}

\item{number_of_points}{Number of points in point cloud.}

\item{domain_size_x}{Size of ROI in x direction.}

\item{domain_size_y}{Size of ROI in y direction.}

\item{search_radius}{Radius to calculate L over.}
}
\value{
L value for each point.
}
\description{
Calculates Getis L values (dev use only).
}
