% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{computeIOU}
\alias{computeIOU}
\title{Compute Intersection Over Union.}
\usage{
computeIOU(vertices_1, vertices_2)
}
\arguments{
\item{vertices_1}{Vertices of first set of hulls. Represented as a matrix with three columns: x, y and vertex label.}

\item{vertices_2}{Vertices of second set of hulls. Represented as a matrix with three columns: x, y and vertex label.}
}
\value{
IOU score.
}
\description{
Takes the labelled vertices of two sets of convex hulls. Calculates the area of intersection over the area of the union of the hulls.
}
