% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIA.R
\name{discretise}
\alias{discretise}
\title{Discretise}
\usage{
discretise(points, splits = NULL, names = NULL, k = 2, createplot = TRUE)
}
\arguments{
\item{points}{Continuous point cloud.}

\item{splits}{Split points for discretisation, will estimate by Gaussian Mixture Modelling if omitted.}
}
\value{
Discrete point cloud.
}
\description{
Converts continuous point cloud into a discrete point cloud.
}
