% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOSEPH.R
\name{filterjoseph}
\alias{filterjoseph}
\title{Filter JOSEPH}
\usage{
filterjoseph(
  josephresults,
  mode = "dispersion",
  minpoints = NULL,
  range = NULL,
  radius = NULL
)
}
\arguments{
\item{josephresults}{Results from JOSEPH.}

\item{mode}{Mode of filtering. Must be a vector containing any combination of the characters "dispersion", "minpoints" and "range".}

\item{minpoints}{Minimum number of points allowed in cluster.}

\item{range}{Range of marked values allowed in cluster.}
}
\value{
Partition of data set.
}
\description{
Filters raw JOSEPH results either by minimum number of points per cluster, range of GP values per cluster or whether dispersion is displayed at the given radius value in Ripley’s H function.
}
