% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{getprobability}
\alias{getprobability}
\title{Get Probability Discrete function used in P-Check.}
\usage{
getprobability(points, inradius, radtotal)
}
\arguments{
\item{points}{Marked point pattern.}

\item{inradius}{Logical matrix specifying if points are in radius of each other.}

\item{radtotal}{Total number of points in radius for each point.}
}
\value{
Probability of non-randomness.
}
\description{
Get Probability Discrete function used in P-Check (dev use only).
}
