% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCalculate.R
\name{gphistogram}
\alias{gphistogram}
\title{GP Histogram}
\usage{
gphistogram(
  gpvalues,
  binwidth = 0.01,
  density = FALSE,
  saveplot = FALSE,
  location = ""
)
}
\arguments{
\item{gpvalues}{A vector of GP values.}
}
\value{
Histogram of GP values.
}
\description{
Creates a histogram of given GP values.
}
