% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIA.R
\name{informclusterpersistence}
\alias{informclusterpersistence}
\title{Inform Cluster Persistence}
\usage{
informclusterpersistence(points, radius = NULL)
}
\arguments{
\item{points}{Point cloud.}

\item{radius}{Radius for determining neighbourhood.}
}
\value{
Persistence Threshold.
}
\description{
Estimates a good persistence threshold for ToMATo clustering.
}
