% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{informdeviance}
\alias{informdeviance}
\title{Inform Deviance.}
\usage{
informdeviance(
  values,
  k = 2,
  createplot = FALSE,
  saveplot = FALSE,
  location = "",
  outputall = FALSE,
  plotresolution = 100
)
}
\arguments{
\item{values}{Vector of marked values from a marked point pattern (e.g. GP values).}

\item{k}{Number of Gaussians to fit. Should be equal to number of suspected categories.}

\item{createplot}{Boolean dictating whether to create a plot of the Gaussian fit.}

\item{saveplot}{Boolean dictating whether to automatically save plot.}

\item{location}{Output location (file path) of saved plot (if any).}

\item{outputall}{Boolean dictating whether to output all results of Gaussian fitting. Mostly used for development.}

\item{plotresolution}{Resolution of x axis used in plot (number of lines).}
}
\value{
Output
}
\description{
Estimates the deviance parameter for JOSEPH clustering by fitting Gaussians to the distribution of marked values.
}
