% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCalculate.R
\name{informradius}
\alias{informradius}
\title{Inform Radius}
\usage{
informradius(points)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
Search radius.
}
\description{
Determine an appropriate search radius based on the argmax of the Ripley’s H function.
}
