% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{intervals}
\alias{intervals}
\title{Intervals}
\usage{
intervals(points, m, n = NULL, xlims = NULL, ylims = NULL)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
m by n intervals.
}
\description{
Used in MPP Segment.
}
