% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOSEPH.R
\name{josephclustersummary}
\alias{josephclustersummary}
\title{JOSEPH Cluster Summary}
\usage{
josephclustersummary(josephresults)
}
\arguments{
\item{josephresults}{Results from JOSEPH.}
}
\value{
Summary statistics of JOSEPH results.
}
\description{
Calculates summary statistics from JOSEPH results.
}
