% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{mom}
\alias{mom}
\title{MPP Histogram.}
\usage{
mom(
  points,
  binsize = 50,
  location = "",
  plotname = "GP Histogram",
  useggplot = FALSE,
  outputdata = FALSE
)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y, and the final column corresponding to mark.}

\item{binsize}{Numeric. Width of bins to use in histogram.}

\item{location}{String or char. Save location (file path) of plots.}

\item{plotname}{String or char. Name of plot for file name.}

\item{useggplot}{Boolean dictating whether to use ggplot to construct histogram.}

\item{outputdata}{Boolean dictating whether to output all results (dev use only).}
}
\value{
A membrane order map.
}
\description{
Creates Membrane Order Map (MOM).
}
