% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPHistogram.R
\name{mpphistogram}
\alias{mpphistogram}
\title{MPP Histogram}
\usage{
mpphistogram(
  points,
  binsize = 50,
  location = "",
  plotname = "GP Histogram",
  useggplot = FALSE,
  outputdata = FALSE
)
}
\arguments{
\item{points}{Point pattern including photon counts in columns 3 and 4.}

\item{binsize}{Size of bins in histogram.}
}
\value{
GP Histogram.
}
\description{
Produces a 2D histogram of GP values.
}
