% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{mppsegment}
\alias{mppsegment}
\title{MPP Segment}
\usage{
mppsegment(points, m, n = NULL, xlims = NULL, ylims = NULL)
}
\arguments{
\item{points}{Point pattern.}

\item{m}{Number of intervals in x direction.}

\item{n}{NUmber of intervals in y direction, will be same as x if omitted.}

\item{xlims}{Maximum x value.}

\item{ylims}{Maximum y value.}
}
\value{
List of point clouds over given intervals.
}
\description{
Segments a large marked point pattern by disregarding all points outside of xlims and ylims (if given) and splitting into m×n equally spaced segments.
}
