% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{photonsimulate}
\alias{photonsimulate}
\title{Simulate Photon Count.}
\usage{
photonsimulate(points, photons)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First three columns corresponding to x, y and GP.}

\item{photons}{Vector of photon counts for each point in points.}
}
\value{
A marked point pattern. Matrix with six rows: x, y, GP, green photon count, red photon count and GP error.
}
\description{
Generates photon counts from given GP values.
}
