% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plothistogramresults}
\alias{plothistogramresults}
\title{Plot Histogram Results function used in PLASMA.}
\usage{
plothistogramresults(i, bindata, GP_histbins, location, points)
}
\arguments{
\item{i}{Integer index.}

\item{bindata}{Data frame or matrix. Output from bindatum().}

\item{GP_histbins}{Data frame or matrix. GP values of bins.}

\item{location}{String or char. Save location (file path).}

\item{points}{Data frame or matrix containing points.}
}
\value{
MPP histogram as per MOM specification.
}
\description{
Plot Histogram Results function used in PLASMA (dev use only).
}
