% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plothulls}
\alias{plothulls}
\title{Plot Hulls.}
\usage{
plothulls(
  points,
  josephresults,
  mingp = NULL,
  maxgp = NULL,
  alpha = 0.5,
  viridis = FALSE,
  saveplot = FALSE,
  location = ""
)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first two columns corresponding to x and y, and the final column corresponding to mark.}

\item{josephresults}{Output from JOSEPH clustering.}

\item{mingp}{Minimum GP value for colour bar. Leave NULL for min in points.}

\item{maxgp}{Maximum GP value for colour bar. Leave NULL for max in points.}

\item{alpha}{Opacity alpha of hulls.}

\item{viridis}{Boolean dictating whether to use viridis colourscale - a cyan to magenta variant will be used if false.}

\item{saveplot}{Boolean dictating whether to automatically save plot.}

\item{location}{Output location (file path) of saved plot (if any).}
}
\value{
Plot of hulls.
}
\description{
Plots convex hulls of clusters found by JOSEPH, coloured by mean GP value.
}
