% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plotmpp}
\alias{plotmpp}
\title{Plot Marked Point Pattern.}
\usage{
plotmpp(points, labels = NULL, range = NULL, includelegend = TRUE)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first three columns corresponding to x, y and mark.}

\item{labels}{Labels of point pattern columns. Will take from points if left NULL.}

\item{range}{Vector of minimum and maximum marked values to define colour bar.}

\item{includelegend}{Boolean dictating whether to show legend (colour bar).}
}
\value{
A plot of a marked point pattern.
}
\description{
Creates a simple plot of a continuous marked point pattern.
}
