% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plotmppdiscrete}
\alias{plotmppdiscrete}
\title{Plot Marked Point Pattern Discrete.}
\usage{
plotmppdiscrete(coords, clusterIndices)
}
\arguments{
\item{coords}{A point pattern. Matrix or data frame with two columns corresponding to x and y.}

\item{clusterIndices}{Cluster indices of each point.}
}
\value{
A plot of the marked point pattern.
}
\description{
Creates a simple plot of a discrete marked point pattern.
}
