% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{ripleyh}
\alias{ripleyh}
\title{Ripley's H Function.}
\usage{
ripleyh(
  points,
  window = NULL,
  createplot = TRUE,
  saveplot = FALSE,
  location = ""
)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y.}

\item{window}{Vector. Bounds of ROI to use, should be written as a vector: (xmin, xmax, ymin, ymax).}

\item{createplot}{Boolean dictating whether to create a plot of the Ripley's H function.}

\item{saveplot}{Boolean dictating whether to automatically save plot.}

\item{location}{String or char. Output location (file path) of saved plot (if any).}
}
\value{
A matrix with two columns: the radii and the corresponding Ripley's H values.
}
\description{
Calculates the Ripley's H function.
}
