% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{saveseg}
\alias{saveseg}
\title{Create and save plot function used in Save Segments.}
\usage{
saveseg(points, loc, titl)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y, and the final column corresponding to mark.}

\item{loc}{String or char. Location to save.}

\item{titl}{String or char. Plot title.}
}
\value{
Saves segmentation.
}
\description{
Create and save plot function used in Save Segments (dev use only).
}
