% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{savesegments}
\alias{savesegments}
\title{Save Segments.}
\usage{
savesegments(segmentation, location)
}
\arguments{
\item{segmentation}{Output of MPP Segment.}
}
\value{
Saves images of segments.
}
\description{
Saves segments found by MPP Segment.
}
