% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{seg}
\alias{seg}
\title{Segment function used in MPP Segment.}
\usage{
seg(x, interval, coords)
}
\arguments{
\item{x}{Integer. x coordinate.}

\item{interval}{Integer. Interval to segment.}

\item{coords}{Data frame or matrix. Coordinates of points.}
}
\value{
Segmented interval.
}
\description{
Segments an MPP (dev use only).
}
