% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{simplehullplot}
\alias{simplehullplot}
\title{Simple Hull Plot.}
\usage{
simplehullplot(
  vertices_1,
  vertices_2,
  legend = FALSE,
  legend_title = " ",
  legend_labels = NULL,
  limits_x = NULL,
  limits_y = NULL
)
}
\arguments{
\item{vertices_1}{Data frame or matrix. Vertices of first set of hulls. Represented as a matrix with three columns: x, y and vertex label.}

\item{vertices_2}{Data frame or matrix. Vertices of second set of hulls. Represented as a matrix with three columns: x, y and vertex label.}

\item{legend}{Boolean dictating whether to show legend in plot.}

\item{legend_title}{String or char which represents the title of the legend (if shown).}

\item{legend_labels}{Vector of strings representing labels of hull types (leave NULL for numeric default).}

\item{limits_x}{Vector representing the minimum and maximum x value of the ROI.}

\item{limits_y}{Vector representing the minimum and maximum y value of the ROI.}
}
\value{
Output
}
\description{
Plots the hulls from two sets of hulls.
}
