% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOSEPH.R
\name{splitjoseph}
\alias{splitjoseph}
\title{Split JOSEPH}
\usage{
splitjoseph(
  josephresults,
  splits = NULL,
  k = 2,
  createplot = FALSE,
  location = "",
  saveplot = FALSE
)
}
\arguments{
\item{josephresults}{Results from JOSEPH.}

\item{splits}{Split points for discretisation, will estimate by Gaussian Mixture Modelling if omitted.}
}
\value{
Partition of data set.
}
\description{
Splits results from JOSEPH into separate data sets based on the splits found by inform persistence. Takes the average label of each cluster and sorts it into one of several categories, then displays each category separately.
}
