% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{toroidaldist}
\alias{toroidaldist}
\title{Toroidal Distance Matrix function used in MPP Simulate.}
\usage{
toroidaldist(points, domain_max_x, domain_max_y)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first two columns corresponding to x and y.}

\item{domain_max_x}{Size of ROI in x direction.}

\item{domain_max_y}{Size of ROI in y direction.}
}
\value{
Toroidal distance matrix.
}
\description{
Calculates toroidal distance matrix of a point cloud.
}
