% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{velocity}
\alias{velocity}
\title{Velocity function used in MPP Simulate.}
\usage{
velocity(
  points,
  number_of_points,
  domain_size_x,
  domain_size_y,
  search_radius,
  D_max,
  L_target
)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first two columns corresponding to x and y.}

\item{number_of_points}{Number of points in point cloud.}

\item{domain_size_x}{Size of ROI in x direction.}

\item{domain_size_y}{Size of ROI in y direction.}

\item{search_radius}{Radius to calculate L over.}

\item{D_max}{Max step size.}

\item{L_target}{Target L value.}
}
\value{
Point offsets.
}
\description{
Determines per-frame velocity of each point (dev use only).
}
