% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{bulkanalyse}
\alias{bulkanalyse}
\title{Bulk Analyse function used in PLASMA.}
\usage{
bulkanalyse(cloud, wholeradius, wholedeviance, split, minpoints)
}
\arguments{
\item{cloud}{Point cloud.}

\item{wholeradius}{Radius used in P-Check and JOSEPH.}

\item{wholedeviance}{Deviance for JOSEPH.}

\item{split}{Threshold for discretising MPP.}

\item{minpoints}{Minimum points per cluster.}
}
\value{
Outputs of each algorithm.
}
\description{
Runs each algorithm used in PLASMA (dev use only).
}
