% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{calculate_L_values}
\alias{calculate_L_values}
\title{Calculate L Values}
\usage{
calculate_L_values(
  distance_matrix,
  number_of_points,
  domain_size_x,
  domain_size_y,
  search_radius
)
}
\arguments{
\item{distance_matrix}{Distance matrix of given point cloud.}

\item{number_of_points}{Number of points.}

\item{domain_size_x}{Domain size in x direction.}

\item{domain_size_y}{Domain size in y direction.}

\item{search_radius}{Search radius defining neighbourhood.}
}
\value{
L Values.
}
\description{
Calculates Getis L Values function used in MPP Simulate.
}
