% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{compilematrix}
\alias{compilematrix}
\title{Compile Matrix function used in MPP Histogram.}
\usage{
compilematrix(i, rep, col)
}
\arguments{
\item{i}{Integer. Bin number.}

\item{rep}{Data frame or matrix. Representative point cloud.}

\item{col}{Integer. Column number.}
}
\value{
Matrix form of bin data.
}
\description{
Converts bindatum output into matrix form. Dev use only.
}
