% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{computeIOUfromjosephs}
\alias{computeIOUfromjosephs}
\title{Compute IOU from JOSEPHs.}
\usage{
computeIOUfromjosephs(josephresults_1, josephresults_2)
}
\arguments{
\item{josephresults_1}{Data frame or matrix. Output from first JOSEPH clustering.}

\item{josephresults_2}{Data frame or matrix. Output from second JOSEPH clustering.}
}
\value{
IOU score.
}
\description{
Computes the IOU between two sets of hulls identified from JOSEPH clustering.
}
