% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{discretise}
\alias{discretise}
\title{Discretise}
\usage{
discretise(points, splits = NULL, names = NULL, k = 2, createplot = TRUE)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First three columns corresponding to x, y and mark.}

\item{splits}{Numeric. Split values which define intervals to categorise points. If left as null, these intervals will be automatically estimated from Gaussian fitting.}

\item{names}{Vector of characters, where each entry corresponds to the name of a category.}

\item{k}{Integer. Number of Gaussians to fit in case splits is left null. Should be equal to number of suspected categories.}

\item{createplot}{Boolean dictating whether to create a plot of the new discretised marked point pattern.}
}
\value{
A discretised marked point pattern.
}
\description{
Converts a continuous marked point pattern into a discrete marked point pattern.
}
