% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLASMA.R
\name{extractresults}
\alias{extractresults}
\title{Extract Results.}
\usage{
extractresults(res, resnum)
}
\arguments{
\item{res}{Results from Bulk Analyse.}
}
\value{
One of the results from P-Check, JOSEPH or MPP Histogram.
}
\description{
Used in PLASMA.
}
