% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{extractresults}
\alias{extractresults}
\title{Extract Results function used in PLASMA.}
\usage{
extractresults(res, resnum)
}
\arguments{
\item{res}{Data frame or matrix. Results from bulkanalyse.}

\item{resnum}{Integer. Results number to extract.}
}
\value{
Algorithm result requested.
}
\description{
Extract Results function used in PLASMA (dev use only).
}
