% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{filterjoseph}
\alias{filterjoseph}
\title{Filter JOSEPH.}
\usage{
filterjoseph(
  josephresults,
  mode = "dispersion",
  minpoints = NULL,
  maxpoints = NULL,
  range = NULL,
  radius = NULL
)
}
\arguments{
\item{josephresults}{Output from JOSEPH clustering.}

\item{mode}{A string representing filter mode. Choose from "dispersion" (filter clusters with above-average dispersion), "minpoints" (removes clusters with point count lower than minpoints), "maxpoints" (removes clusters with point count higher than maxpoints) and "range" (removes clusters with mean marked value outside the range).}

\item{minpoints}{Minimum number of points.}

\item{maxpoints}{Maximum number of points.}

\item{range}{Vector representing lower and upper bounds of average mark to include.}

\item{radius}{Radius of clusters used to check dispersion.}
}
\value{
Filtered JOSEPH results. The input marked point pattern with updated cluster allocation.
}
\description{
Filters JOSEPH cluster results based on user requirements.
}
