% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIA.R
\name{getari}
\alias{getari}
\title{Get Adjusted Rand Index}
\usage{
getari(
  points,
  clusterpersistence = NULL,
  clusterradius = NULL,
  createplot = FALSE,
  location = "",
  plotname = "",
  clusterwhole = NULL
)
}
\arguments{
\item{points}{Point pattern.}

\item{clusterradius}{Radius for determining neighbourhood.}

\item{clusterperistence}{Persistence threshold for ToMATo.}
}
\value{
Adjusted rand index.
}
\description{
Allows user to determine ARI value associated with the whole and individual partitions of the data set.
}
