% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{getari}
\alias{getari}
\title{Get ARI (Proxy for ToMATo Clustering)}
\usage{
getari(points, search_radius, persistence)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y.}

\item{search_radius}{Numeric. Cluster radius.}

\item{persistence}{Numeric. Persistence threshold.}
}
\value{
Point cloud with cluster index.
}
\description{
Perform ToMATo clustering.
}
