% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROBABILITY.R
\name{getprobability}
\alias{getprobability}
\title{Get Probability}
\usage{
getprobability(points, inradius, radtotal)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
Weighted proportion.
}
\description{
Determines the weighted proportion of neighbouring points with the same category as the root points.
}
