% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCalculate.R
\name{gpcalculate}
\alias{gpcalculate}
\title{GP Calculate}
\usage{
gpcalculate(points)
}
\arguments{
\item{points}{Point pattern including photon counts in columns 3 and 4.}
}
\value{
Vector of GP values.
}
\description{
Calculates the GP values of a point cloud given the photon counts.
}
