% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{gpcalculate}
\alias{gpcalculate}
\title{GP Calculate}
\usage{
gpcalculate(points)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with third and fourth columns corresponding to green and red photon counts.}
}
\value{
Vector of GP values of each localisation.
}
\description{
Calculate the GP values from a marked point pattern.
}
