% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{gphistogram}
\alias{gphistogram}
\title{GP Histogram}
\usage{
gphistogram(
  gpvalues,
  binwidth = 0.01,
  density = FALSE,
  saveplot = FALSE,
  location = ""
)
}
\arguments{
\item{gpvalues}{Vector of GP values.}

\item{binwidth}{Numeric. Width of bin used in histogram.}

\item{density}{Boolean dictating whether to slow density plot overlaid on histogram.}

\item{saveplot}{Boolean dictating whether to automatically save plot.}

\item{location}{String or char. Output location (file path) of saved plot (if any).}
}
\value{
Plot of GP values.
}
\description{
Create histogram of all GP values.
}
