% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{informclusterpersistence}
\alias{informclusterpersistence}
\title{Inform Cluster Persistence}
\usage{
informclusterpersistence(points, radius = NULL)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y.}

\item{radius}{Numeric. Suspected cluster radius. If left null, this will be estimated from the Ripley's function.}
}
\value{
A persistence threshold.
}
\description{
Estimates the persistence threshold to use for ToMATo clustering.
}
