% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{informradius}
\alias{informradius}
\title{Inform Radius}
\usage{
informradius(points)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y.}
}
\value{
An estimate of the cluster radius.
}
\description{
Estimate the average cluster radius of a point cloud using the Ripley's H function.
}
