% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{intervals}
\alias{intervals}
\title{Intervals function used in MPP Segment.}
\usage{
intervals(points, m, n = NULL, xlims = NULL, ylims = NULL)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first two columns corresponding to x and y, and the final column corresponding to mark.}

\item{m}{Number of x intervals.}

\item{n}{Number of y intervals. Will use m if left NULL.}

\item{xlims}{Vector of minimum and maximum x values to use.}

\item{ylims}{Vector of minimum and maximum y values to use.}
}
\value{
Segmented intervals.
}
\description{
Segments an MPP (dev use only).
}
