% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{joseph}
\alias{joseph}
\title{Justification of Separation by Employed Persistent Homology (JOSEPH)}
\usage{
joseph(points, deviance = NULL, radius = NULL)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first two columns corresponding to x and y, and the final column corresponding to mark.}

\item{deviance}{Deviance parameter - acceptable deviation from mean mark value of cluster.}

\item{radius}{Cluster radius.}
}
\value{
The input marked point pattern with an additional column representing cluster allocation.
}
\description{
A topological data analysis tool for clustering marked point pattern data.
}
