% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOSEPH.R
\name{joseph}
\alias{joseph}
\title{JOSEPH}
\usage{
joseph(points, deviance = NULL, radius = NULL)
}
\arguments{
\item{points}{Numeric point pattern.}

\item{deviance}{Maximum difference in marked values between roots of clusters and any other point.}

\item{radius}{Search radius defining neighbourhood of each point.}
}
\value{
Partition of data set.
}
\description{
Outputs a partition of the data set, clustering by distance and marks.
}
