% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{josephclustersummary}
\alias{josephclustersummary}
\title{JOSEPH Cluster Summary function used in PLASMA.}
\usage{
josephclustersummary(josephresults)
}
\arguments{
\item{josephresults}{Output from JOSEPH clustering.}
}
\value{
A range of cluster statistics, including number of points, area of convex hull, density, average mark and mark standard deviation.
}
\description{
Outputs cluster statistics from JOSEPH results.
}
