% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{mppsimulate}
\alias{mppsimulate}
\title{MPP Simulate.}
\usage{
mppsimulate(
  number_of_frames = 100,
  number_of_points = 500,
  domain_size_x = 1000,
  domain_size_y = NULL,
  search_radius = 50,
  L_target = 100,
  D_max = 50,
  tomato_threshold = NULL,
  grid_size = 100,
  number_of_categories = 1,
  number_of_resimulated_points = NULL,
  use_csr = FALSE,
  probability_of_picking_categories = c(0.7),
  precision_alpha = 0,
  precision_beta = 0,
  category_mean = c(1),
  category_sd = c(0),
  background_mean = 0,
  background_sd = 0,
  createplots = FALSE,
  saveplots = FALSE,
  location = "",
  plotname = "Plot",
  heatmapname = "Heatmap",
  outputall = FALSE
)
}
\arguments{
\item{number_of_frames}{Number of frames to simulate over.}

\item{number_of_points}{Number of points to use in simulation.}

\item{domain_size_x}{Size of ROI in x direction.}

\item{domain_size_y}{Size of ROI in y direction. Will be same as x if left NULL.}

\item{search_radius}{Radius to calculate Getis L value.}

\item{L_target}{Target L value to fit to.}

\item{D_max}{Max step size between frames.}

\item{tomato_threshold}{Persistence threshold for ToMATo clustering when producing spatial clusters. Leave NULL for automatic estimation.}

\item{grid_size}{Granularity of grid used to place marked points overlaid on spatial clusters.}

\item{number_of_categories}{Number of marked point types, each with own Normal distribution of marked values.}

\item{number_of_resimulated_points}{Number of points to use in marked point pattern, leave as NULL for same as number_of_points.}

\item{use_csr}{Boolean dictating whether to use a completely spatially random distribution for the marked point pattern.}

\item{probability_of_picking_categories}{Probability of assigning each spatial cluster to a given category.}

\item{precision_alpha}{Alpha parameter of the Gamma distribution used to simulate localisation error.}

\item{precision_beta}{Beta parameter of the Gamma distribution used to simulate localisation error.}

\item{category_mean}{Vector of mean values of the Normal distributions used to generate marks for each category.}

\item{category_sd}{Vector of standard deviations of the Normal distributions used to generate marks for each category.}

\item{background_mean}{Mean value of the Normal distribution used to generate marks for the background points.}

\item{background_sd}{Standard deviation of the Normal distribution used to generate marks for the background points.}

\item{createplots}{Boolean dictating whether to create plots of the marked point pattern and hull heatmap.}

\item{saveplots}{Boolean dictating whether to automatically save plots.}

\item{location}{Output location (file path) of saved plots (if any).}

\item{plotname}{Name of plot for file name.}

\item{heatmapname}{Name of heatmap for file name.}

\item{outputall}{Boolean dictating whether all results should be output (dev use only).}
}
\value{
A simulated marked point pattern. A matrix with 4 columns: x, y, assigned category and marked value.
}
\description{
Versatile simulator for generating marked point patterns with pre-defined properties. Produces spatial clusters by trying to fit to a given Getis L value over a given radius. These are then overlaid with marked point clusters. Each marked point is randomly assigned a category, with corresponding Gaussian distribution for generating marked values.
}
