% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{pcheck}
\alias{pcheck}
\title{Probability Check function.}
\usage{
pcheck(points, number_of_trials = 100, radius = NULL, outputall = FALSE)
}
\arguments{
\item{points}{A marked point pattern. Matrix or data frame with first three columns corresponding to x, y and discretised mark.}

\item{number_of_trials}{Number of permutations to run.}

\item{radius}{Suspected domain radius (will be estimated if left NULL).}

\item{outputall}{Output all results (dev use only).}
}
\value{
P-value.
}
\description{
Performs P-Check, which determines the p-value of non-random partitioning of points by spatial and discrete marked values (probability of domains appearing by random chance).
}
