% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROBABILITY.R
\name{pcheck}
\alias{pcheck}
\title{P Check}
\usage{
pcheck(points, number_of_trials = 100, radius = NULL, outputall = FALSE)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
Weighted proportion and associated p value.
}
\description{
Determines the weighted proportion of a point pattern and outputs a p-value associated with in.
}
