% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{photonsimulate}
\alias{photonsimulate}
\title{Photon Simulate}
\usage{
photonsimulate(points, photons)
}
\arguments{
\item{points}{Point pattern (including GP values in third column).}

\item{photons}{Vector if photon counts in first channel.}
}
\value{
Point pattern with photon counts and GP values.
}
\description{
Adds additional photon count estimations to a given simulated marked point pattern (if it does not have them already).
}
