% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLASMA.R
\name{plasma}
\alias{plasma}
\title{PLASMA}
\usage{
plasma(
  points,
  location = "",
  radius = NULL,
  deviance = NULL,
  split = NULL,
  minpoints = 2,
  names = NULL
)
}
\arguments{
\item{points}{Point pattern or list of point patterns.}

\item{location}{Output location for results.}
}
\value{
Results of P-Check, JOSEPH and MPP Histogram, including: histogram of GP values with density plot, density plots of fitted Gaussians with all Gaussian parameters, Ripley’s H Function, all suggested parameters, all P-Check p-values, all JOSEPH results (points data), a plot of all JOSEPH clusters, summary statistics of clusters found by JOSEPH over all clouds (points per cluster, area of convex/concave hull of cluster, density, mean mark per cluster, variance of mark values per cluster, global mean mark, difference between global mean and cluster mean), all MPP Histogram results.
}
\description{
Performs P-Check, JOSEPH and MPP Histogram on a point pattern or list of point patterns.
}
