% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plasma}
\alias{plasma}
\title{Point Label Analysis of Super-resolved Membrane Attributes}
\usage{
plasma(
  points,
  location = "",
  radius = NULL,
  deviance = NULL,
  split = NULL,
  minpoints = 2,
  names = NULL
)
}
\arguments{
\item{points}{Data frame, matrix or list. A point cloud or list of point clouds (each defined as above).}

\item{location}{String or char. Output location (file path) for all results to be saved.}

\item{radius}{Numeric. Cluster radius to use for P-Check and JOSEPH. Will be estimated if left NULL.}

\item{deviance}{Numeric. Deviance to use for JOSEPH. Will be estimated if left NULL.}

\item{split}{Numeric. Split points to discretise marked point pattern. Will be estimated from Gaussian fitting if left NULL.}

\item{minpoints}{Integer. Minimum number of points per cluster.}

\item{names}{Vector of names associated with each point pattern. Will use numeric if left NULL.}
}
\value{
All outputs of P-Check, JOSEPH and MOM, saved at given location.
}
\description{
Performs P-Check, JOSEPH and MOM. All data should be presented as a data frame with four or five columns: x coordinates, y coordinates, photon count first channel, photon count second channel and (optional, will be calculated if not given) GP. Can input either a single point cloud or a list of point clouds. Each point cloud should have undergone suitable pre-processing (i.e. only include regions) within the cell.
}
