% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLASMA.R
\name{plothistogramresults}
\alias{plothistogramresults}
\title{Plot Histogram Results}
\usage{
plothistogramresults(i, bindata, GP_histbins, location, points)
}
\arguments{
\item{i}{Index.}
}
\value{
Plots of results from MPP Histogram.
}
\description{
Used in PLASMA.
}
