% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{plothistogramresults}
\alias{plothistogramresults}
\title{Plot Histogram Results function used in PLASMA.}
\usage{
plothistogramresults(i, bindata, GP_histbins, location, points)
}
\arguments{
\item{i}{Index.}

\item{bindata}{Output from bindatum().}

\item{GP_histbins}{GP values of bins.}

\item{location}{Save location (file path).}

\item{points}{Point cloud.}
}
\value{
MPP histogram as per MOM specification.
}
\description{
Plot Histogram Results function used in PLASMA (dev use only).
}
