% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{plotmpp}
\alias{plotmpp}
\title{Plot MPP}
\usage{
plotmpp(points, labels = NULL, range = NULL, includelegend = TRUE)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
Plot of points.
}
\description{
Plots a continuous marked point pattern.
}
