% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{plotmppdiscrete}
\alias{plotmppdiscrete}
\title{Plot MPP Discrete}
\usage{
plotmppdiscrete(coords, clusterIndices)
}
\arguments{
\item{coords}{Point pattern.}

\item{clusterIndices}{Labels or cluster indices of points.}
}
\value{
Plot of points.
}
\description{
Plots a discrete marked point pattern, based on plotClusterScatter from RSMLM.
}
