% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCalculate.R
\name{ripleyh}
\alias{ripleyh}
\title{Ripley's H Function}
\usage{
ripleyh(
  points,
  window = NULL,
  createplot = TRUE,
  saveplot = FALSE,
  location = ""
)
}
\arguments{
\item{points}{Point pattern.}
}
\value{
Plot of Ripley's H Function.
}
\description{
Calculates Ripley’s H function using the spatstat Ripley’s K Function with isotropic edge correction and gives a line plot.
}
