% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{savesegments}
\alias{savesegments}
\title{Save Segments.}
\usage{
savesegments(segmentation, location)
}
\arguments{
\item{segmentation}{Output of mppsegment().}

\item{location}{Save location (file path).}
}
\value{
Plot of each ROI.
}
\description{
Saves each segmented ROI output from mppsegment().
}
