% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{splitjoseph}
\alias{splitjoseph}
\title{Split JOSEPH.}
\usage{
splitjoseph(
  josephresults,
  splits = NULL,
  k = 2,
  createplot = FALSE,
  location = "",
  saveplot = FALSE
)
}
\arguments{
\item{josephresults}{Data frame or matrix. Output from JOSEPH clustering.}

\item{splits}{Numeric. Split values which define intervals to categorise points. If left as null, these intervals will be automatically estimated from Gaussian fitting.}

\item{k}{Integer. Number of Gaussians to fit in case splits is left null. Should be equal to number of suspected categories.}

\item{createplot}{Boolean dictating whether to create a plot of the new discretised marked point pattern.}

\item{location}{String or char. Output location (file path) of saved plot (if any).}

\item{saveplot}{Boolean dictating whether to automatically save plot.}
}
\value{
A discretised marked point pattern.
}
\description{
Discretises JOSEPH results as in discretise(), with each cluster assigned a different category.
}
