% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{toroidaldist}
\alias{toroidaldist}
\title{Toroidal Distance}
\usage{
toroidaldist(points, domain_max_x, domain_max_y)
}
\arguments{
\item{points}{Point cloud.}

\item{domain_max_x}{Domain size in x direction.}

\item{domain_max_y}{Domain size in y direction.}
}
\value{
Toroidal distance matrix.
}
\description{
Calculates distance between points on toroidal coordinate system.
}
