% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPPSimulate.R
\name{velocity}
\alias{velocity}
\title{Velocity}
\usage{
velocity(
  points,
  number_of_points,
  domain_size_x,
  domain_size_y,
  search_radius,
  D_max,
  L_target
)
}
\arguments{
\item{points}{Point cloud.}

\item{number_of_points}{Number of points.}

\item{domain_size_x}{Domain size in x direction.}

\item{domain_size_y}{Domain size in y direction.}

\item{search_radius}{Search radius defining neighbourhood.}

\item{D_max}{Diffusivity multiplier for offset.}

\item{L_target}{Target L value.}
}
\value{
Offset.
}
\description{
Velocity function used to determine offset of randomly generated points in MPP Simulate.
}
