% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfunctions.R
\name{velocity}
\alias{velocity}
\title{Velocity function used in MPP Simulate.}
\usage{
velocity(
  points,
  number_of_points,
  domain_size_x,
  domain_size_y,
  search_radius,
  D_max,
  L_target
)
}
\arguments{
\item{points}{Data frame or matrix. A marked point pattern. First two columns corresponding to x and y.}

\item{number_of_points}{Numeric. Number of points in point cloud.}

\item{domain_size_x}{Numeric. Size of ROI in x direction.}

\item{domain_size_y}{Numeric. Size of ROI in y direction.}

\item{search_radius}{Numeric. Radius to calculate L over.}

\item{D_max}{Numeric. Max step size.}

\item{L_target}{Numeric. Target L value.}
}
\value{
Point offsets.
}
\description{
Determines per-frame velocity of each point (dev use only).
}
